Backend/requirements.txt
python-dotenv
google-generativeai
requests
fastapi
pydantic
langgraph
asyncio
fastapi==0.110
tavily-python
uvicorn[standard]==0.29
python-dotenv==1.0
openai==1.30
pydantic==2.7
httpx==0.27
apify-client
groq